requirements.txt
anthropic>=0.40.0
elevenlabs>=1.0.0
watchdog>=4.0.0
APScheduler>=3.10.0
pdfminer.six>=20221105
python-dotenv>=1.0.0
PyYAML>=6.0
loguru>=0.7.0
mutagen>=1.47.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
